Refactor loading indicator logic
