+++ intermine-ios/LoadingTableViewController.swift
@@ -1,0 +1,53 @@
+//
+//  LoadingTableViewController.swift
+//  intermine-ios
+//
+//  Created by Nadia on 5/12/17.
+//  Copyright © 2017 Nadia. All rights reserved.
+//
+
+import UIKit
+import NVActivityIndicatorView
+
+class LoadingTableViewController: UITableViewController {
+    
+    var spinner: NVActivityIndicatorView?
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        self.spinner = NVActivityIndicatorView(frame: self.indicatorFrame(), type: .ballGridPulse, color: Colors.chelseaCucumber, padding: self.indicatorPadding())
+        if let spinner = self.spinner {
+            self.view.addSubview(spinner)
+            self.view.bringSubview(toFront: spinner)
+        }
+        self.spinner?.startAnimating()
+    }
+    
+    private func indicatorFrame() -> CGRect {
+        if let navbarHeight = self.navigationController?.navigationBar.frame.size.height, let tabbarHeight = self.tabBarController?.tabBar.frame.size.height {
+            let viewHeight = BaseView.viewHeight(view: self.view)
+            let indicatorHeight = viewHeight - (tabbarHeight + navbarHeight)
+            let indicatorWidth = BaseView.viewWidth(view: self.view)
+            return CGRect(x: 0, y: 0, width: indicatorWidth, height: indicatorHeight)
+        } else {
+            return self.view.frame
+        }
+    }
+    
+    private func indicatorPadding() -> CGFloat {
+        return BaseView.viewWidth(view: self.view) / 2.5
+    }
+
+
+    // MARK: - Table view data source
+
+    override func numberOfSections(in tableView: UITableView) -> Int {
+        // To override
+        return 0
+    }
+
+    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        // To override
+        return 0
+    }
+}
